move classes related to crac/cnec and xnode from the tt_res package to cnecs and xnode packages to use them in the ttc rao for better organization
