+++ cse-lib/data/src/main/java/com/farao_community/farao/cse/data/cnec/FlowCnecResult.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.farao_community.farao.cse.data.ttc_res;
+package com.farao_community.farao.cse.data.cnec;
 
 /**
  * @author Joris Mancini {@literal <joris.mancini at rte-france.com>}
