+++ cse-lib/data/src/main/java/com/farao_community/farao/cse/data/xnode/XNodeReader.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.farao_community.farao.cse.data.ttc_res;
+package com.farao_community.farao.cse.data.xnode;
 
 import com.farao_community.farao.cse.data.CseDataException;
 import com.farao_community.farao.cse.data.DataUtil;
