+++ cse-lib/data/src/test/java/com/farao_community/farao/cse/data/ttc_res/TtcResultTest.java
@@ -7,6 +7,7 @@
 
 package com.farao_community.farao.cse.data.ttc_res;
 
+import com.farao_community.farao.cse.data.cnec.CracResultsHelper;
 import com.farao_community.farao.cse.data.xsd.ttc_res.Timestamp;
 import com.farao_community.farao.data.crac_api.Crac;
 import com.farao_community.farao.data.crac_io_api.CracImporters;
