--- cse-lib/data/src/test/resources/com/farao_community/farao/cse/data/ttc_res/fake_xnodes.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0"?>
-<xnodes xmlns="http://www.rte-france.com/gsr"
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-    <xnodeTag name="XBI_MO31" area1="AL" area2="GR" />
-
-</xnodes>
